refactor(client): split conversation_panel into parts
